requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/LLM
google-generativeai==0.3.2
ollama==0.1.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Email Processing
email-validator==2.1.0
python-multipart==0.0.6

# File Processing
python-magic-bin==0.4.14  # Windows
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
